add two bv moran's I nodes
